Extract ControllerRadioButtons to a separate file
